Bugfix: Nope - that didn't work, code functioned but then quantity appeared as 0 for values between greater han 0 but less than one (e.g. items that are sold in weight quantities). Have changed data format from integer to float across board. It now shows everything in qty field with a decimal...but it works!

+++ app/src/main/java/com/stevecrossin/grocerytracker/adapters/ReceiptItemsAdapter.java
@@ -42,7 +42,7 @@ public class ReceiptItemsAdapter extends RecyclerView.Adapter<ReceiptItemsAdapte
         holder.mTextViewUnitPriceLabel.setText(mHeader.unitPriceHeader);
         holder.mTextViewUnitPriceValue.setText(String.format("$%s", decimalFormat.format(receiptLineItem.unitPrice)));
         holder.mTextViewQuantityLabel.setText(mHeader.quantityHeader);
-        holder.mTextViewQuantityValue.setText(String.format(Integer.toString(receiptLineItem.quantity)));
+        holder.mTextViewQuantityValue.setText(String.format(Float.toString(receiptLineItem.quantity)));
         holder.mTextViewPriceLabel.setText(mHeader.priceHeader);
         holder.mTextViewPriceValue.setText(String.format("$%s", decimalFormat.format(receiptLineItem.price)));
     }
